# docs/swagger.yaml
basePath: /api/v1
definitions:
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  model.Ip:
    properties:
      IP:
        description: json 回傳時IP 的 Key 對應到 IP
        type: string
      時間:
        description: json 回傳時Time 的 Key 對應到 時間
        type: string
      狀態:
        type: number
      網址:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server mysql server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /iplogs:
    get:
      consumes:
      - application/json
      description: get string by IP
      parameters:
      - description: iplogs search by q
        format: email
        in: query
        name: ip
        type: string
      produces:
      - application/json
      responses: {}
      summary: Show a IP
      tags:
      - iplogs
    post:
      consumes:
      - application/json
      description: add by json body
      parameters:
      - description: Add iplogs
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/model.Ip'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Add an Ip
      tags:
      - iplogs
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"